requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
langchain==0.2.6
langchain-community==0.2.6
langchain-core==0.2.10
langchain-huggingface>=0.0.3
langchain-text-splitters>=0.2.0
faiss-cpu
rank_bm25
sentence-transformers
docling
pypdfium2
Pillow
easyocr
pythainlp
python-multipart
httpx
pydantic>=2.0
numpy<2.0.0
python-dotenv